+++ examples/src/bin/ieee802154_sniffer.rs
@@ -10,6 +10,8 @@
 use esp_backtrace as _;
 use esp_hal::{
     clock::ClockControl,
+    default_uart0_pins,
+    gpio::Io,
     peripherals::Peripherals,
     prelude::*,
     reset::software_reset,
@@ -25,7 +27,13 @@ fn main() -> ! {
     let system = SystemControl::new(peripherals.SYSTEM);
     let clocks = ClockControl::max(system.clock_control).freeze();
 
-    let mut uart0 = Uart::new(peripherals.UART0, &clocks);
+    let io = Io::new(peripherals.GPIO, peripherals.IO_MUX);
+
+    // Default pins for Uart/Serial communication
+    let (mut tx_pin, mut rx_pin) = default_uart0_pins!(io);
+
+    let mut uart0 =
+        Uart::new_with_default_pins(peripherals.UART0, &clocks, &mut tx_pin, &mut rx_pin).unwrap();
 
     // read two characters which get parsed as the channel
     let mut cnt = 0;